+++ src/cutadapt/__init__.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import print_function, division, absolute_import
-
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions