+++ src/cutadapt/align.py
@@ -1,9 +1,6 @@
-# coding: utf-8
 """
 Alignment module.
 """
-from __future__ import print_function, division, absolute_import
-
 from cutadapt._align import Aligner, compare_prefixes, locate
 
 # flags for global alignment