+++ src/cutadapt/colorspace.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 """
 Colorspace conversion routines.
 
@@ -13,9 +12,6 @@ C 1 0 3 2
 G 2 3 0 1
 T 3 2 1 0
 """
-from __future__ import print_function, division, absolute_import
-
-__author__ = 'Marcel Martin'
 
 
 def _initialize_dicts():