+++ src/cutadapt/modifiers.py
@@ -1,15 +1,12 @@
-# coding: utf-8
 """
 This module implements all the read modifications that cutadapt supports.
 A modifier must be callable. It is implemented as a function if no parameters
 need to be stored, and as a class with a __call__ method if there are parameters
 (or statistics).
 """
-from __future__ import print_function, division, absolute_import
 import re
 from collections import OrderedDict
 from cutadapt.qualtrim import quality_trim_index, nextseq_trim_index
-from cutadapt.compat import maketrans
 
 
 class AdapterCutter(object):
@@ -176,7 +173,7 @@ class DoubleEncoder(object):
 	Double-encode colorspace reads, using characters ACGTN to represent colors.
 	"""
 	def __init__(self):
-		self.double_encode_trans = maketrans('0123.', 'ACGTN')
+		self.double_encode_trans = str.maketrans('0123.', 'ACGTN')
 
 	def __call__(self, read, matches):
 		read = read[:]
@@ -190,7 +187,7 @@ class ZeroCapper(object):
 	"""
 	def __init__(self, quality_base=33):
 		qb = quality_base
-		self.zero_cap_trans = maketrans(''.join(map(chr, range(qb))), chr(qb) * qb)
+		self.zero_cap_trans = str.maketrans(''.join(map(chr, range(qb))), chr(qb) * qb)
 
 	def __call__(self, read, matches):
 		read = read[:]