+++ tests/test_colorspace.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import print_function, division, absolute_import
-
 from cutadapt.colorspace import encode, decode
 from cutadapt.__main__ import main
 from utils import run, datapath