+++ tests/test_modifiers.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import print_function, division, absolute_import
-
 from cutadapt.seqio import Sequence
 from cutadapt.modifiers import (UnconditionalCutter, NEndTrimmer, QualityTrimmer,
 	Shortener)