+++ tests/test_qualtrim.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-from __future__ import print_function, division, absolute_import
-
 from cutadapt.seqio import Sequence
 from cutadapt.qualtrim import nextseq_trim_index